+++ CHANGELOG.rst
@@ -4,6 +4,7 @@
 * Added support for returning multiple objects. (`#7`_)
 * Refactored models instantiation. (`#9`_)
 * Added `Conversion` model. (`#3`_)
+* Better models representation. (`#8`_)
 
 0.1.2
 =====
@@ -22,6 +23,7 @@
 
 
 .. _#9: https://github.com/Stranger6667/pyoffers/issues/9
+.. _#8: https://github.com/Stranger6667/pyoffers/issues/8
 .. _#7: https://github.com/Stranger6667/pyoffers/issues/7
 .. _#5: https://github.com/Stranger6667/pyoffers/issues/5
 .. _#3: https://github.com/Stranger6667/pyoffers/issues/3